refactor: single swap for reconcile, similar to queue implementation options
